refactor(core): CountNonZero and ComponentMap::num_labels use std::size_t
Change CountNonZero::operator() return type from int to std::size_t,
and ComponentMap::num_labels / count() from int to std::size_t.
Update check() to avoid signed/unsigned comparison warnings.
Update test assertions to use unsigned literal suffixes.

+++ tests/core/ops/test_connected_components.cpp
@@ -73,7 +73,7 @@ TEST(ComponentMapTest, MaskIsCorrectType) {
 TEST(ComponentMapTest, OutOfRangeLabelThrows) {
     Image<Gray> src = make_one_blob();
     ComponentMap result = src | ConnectedComponents{};
-    EXPECT_THROW(result.area(result.count()), std::out_of_range);
+    EXPECT_THROW(result.area(static_cast<int>(result.count())), std::out_of_range);
     EXPECT_THROW(result.area(-1), std::out_of_range);
 }
 
@@ -88,7 +88,7 @@ TEST(ConnectedComponentsTest, FindsComponentsInBinaryImage) {
 TEST(ConnectedComponentsTest, BlackImageHasOnlyBackground) {
     Image<Gray> src(cv::Mat(50, 50, CV_8UC1, cv::Scalar(0)));
     ComponentMap result = src | ConnectedComponents{};
-    EXPECT_EQ(result.count(), 1);
+    EXPECT_EQ(result.count(), 1u);
 }
 
 TEST(ConnectedComponentsTest, PipelineSyntax) {
